ETL Project/requirements.txt
spotipy
pandas
apache-airflow-providers-postgres
matplotlib
SQLAlchemy
psycopg2-binary
requests
apache-airflow
datetime
